Tengo demasiado sueño, cambie algunas cosas

+++ back/src/main/java/com/uade/back/dto/catalog/CategoryIdRequest.java
@@ -1,2 +1,2 @@
 package com.uade.back.dto.catalog;
-public record CategoryIdRequest(Long id) {}
+public record CategoryIdRequest(Integer id) {}
