+++ back/src/main/java/com/uade/back/dto/image/ImageIdRequest.java
@@ -1,2 +1,2 @@
 package com.uade.back.dto.image;
-public record ImageIdRequest(Long id) {}
+public record ImageIdRequest(Integer id) {}
